[Refactor] merge split files into a single file
