+++ src/main/java/com/capston/favicon/application/UserServiceImpl.java
@@ -3,8 +3,12 @@ package com.capston.favicon.application;
 import com.capston.favicon.application.repository.UserService;
 import com.capston.favicon.domain.domain.User;
 import com.capston.favicon.infrastructure.UserRepository;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 
@@ -22,6 +26,27 @@ public class UserServiceImpl implements UserService {
             throw new IllegalArgumentException("아이디가 이미 존재합니다. 재확인해주세요.");
         }
         userRepository.save(user);
+    }
+
+    @Override
+    public void login(String username, String password, HttpServletRequest request) {
+        User user = userRepository.findByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException(username);
+        }
+        if (user.getPassword().equals(password)) {
+            HttpSession session = request.getSession();
+            session.setAttribute("username", username);
+        } else {
+            throw new BadCredentialsException("Wrong password");
+        }
+
+    }
+
+    @Override
+    public void logout(HttpServletRequest request) {
+        HttpSession session = request.getSession();
+        session.removeAttribute("username");
     }
 
 }
