--- src/main/java/com/capston/favicon/application/repository/AuthService.java
@@ -1,8 +1,0 @@
-package com.capston.favicon.application.repository;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public interface AuthService {
-    void login(String username, String password, HttpServletRequest request);
-    void logout(HttpServletRequest request);
-}
