+++ src/main/java/com/capston/favicon/application/repository/UserService.java
@@ -1,7 +1,10 @@
 package com.capston.favicon.application.repository;
 
 import com.capston.favicon.domain.domain.User;
+import jakarta.servlet.http.HttpServletRequest;
 
 public interface UserService {
     void join(User user);
+    void login(String username, String password, HttpServletRequest request);
+    void logout(HttpServletRequest request);
 }
